Extract IR-related virtual files in separate files.
This commit is just cut-and-paste. The intent is to be able to move
the extracted files in separate artifacts.
