+++ tools/js/src/main/scala/org/scalajs/core/tools/io/NodeVirtualIRFiles.scala
@@ -1,0 +1,12 @@
+/*                     __                                               *\
+**     ________ ___   / /  ___      __ ____  Scala.js tools             **
+**    / __/ __// _ | / /  / _ | __ / // __/  (c) 2013-2018, LAMP/EPFL   **
+**  __\ \/ /__/ __ |/ /__/ __ |/_// /_\ \    http://scala-js.org/       **
+** /____/\___/_/ |_/____/_/ | |__/ /____/                               **
+**                          |/____/                                     **
+\*                                                                      */
+
+package org.scalajs.core.tools.io
+
+class NodeVirtualScalaJSIRFile(p: String)
+    extends NodeVirtualBinaryFile(p) with VirtualSerializedScalaJSIRFile
